Issue #no-48 textarea field added
